fixed and refactored: - types - globals - hooks
