--- src/types/CardData.tsx
@@ -1,6 +1,0 @@
-export type CardData = {
-    title: string;
-    description: string;
-    price: number;
-    custom: boolean;
-}
