--- src/types/CustomDataProps.tsx
@@ -1,3 +1,0 @@
-export interface CustomDataProps {
-    customElement: string;
-}
