+++ src/types/types.tsx
@@ -1,0 +1,11 @@
+export interface CustomDataType {
+    customElement: string;
+}
+
+export type CardDataType = {
+    title: string;
+    description: string;
+    price: number;
+    custom: boolean;
+}
+
